refactored assetbuilder - mock tests?

+++ src/Build/AssetBuilder.php
@@ -12,8 +12,6 @@ use Assetic\Cache\FilesystemCache;
 
 class AssetBuilder implements BuilderInterface {
   private $buildContext;
-  private $siteJsFile;
-  private $siteCssFile;
   private $cssPath = 'css/';
   private $jsPath = 'js/';
   private $imagesPath = 'images/';
@@ -30,17 +28,6 @@ class AssetBuilder implements BuilderInterface {
     if(isset($options['cssPath'])) $this->cssPath = $options['cssPath'];
     if(isset($options['imagesPath'])) $this->imagesPath = $options['imagesPath'];
 
-    /*$this->jsPath = $this->buildContext->getPath('assetJs');
-    $this->cssPath = $this->buildContext->getPath('assetCss');
-    $this->imagesPath = $this->buildContext->getPath('assetImages');
-    */
-  }
-
-  private function setSiteJsFile(string $buildId = ""){
-    $this->siteJsFile = $this->jsPath . "/$buildId-site.js";
-  }
-  private function setSiteCssFile(string $buildId = ""){
-    $this->siteCssFile = $this->cssPath . "/$buildId-site.css";
   }
 
   public function setCompression(string $compression){
@@ -57,54 +44,40 @@ class AssetBuilder implements BuilderInterface {
   }
 
   public function beforeBuild(){
-    //clear out any css or js files in build
-    $path = $this->buildContext->getPath('build') . '/' . $this->sourceContext;
-    $files = array_merge(
-      glob($path . $this->cssPath . "/*"),
-      glob($path . $this->jsPath . "/*")
-    );
-    foreach($files as $file){
-      unlink($file);
+    //clear out any css or js files in build - if production
+    if($this->buildContext->getEnv() == 'production'){
+      foreach($this->getBuiltAssets() as $file){
+        unlink($file);
+      }
     }
   }
 
-  public function build(array $buildArgs = []){
+  public function build(array $buildArgs = []) :array{
     if(!isset($buildArgs['buildID'])){
       $buildArgs['buildID'] = substr(sha1(microtime()), 0, 7);
     }
-    $this->setSiteJsFile($buildArgs['buildID']);
-    $this->setSiteCssFile($buildArgs['buildID']);
 
     if($this->buildContext->getEnv() == 'production') {
-      $this->buildJsAssets();
-      $this->buildCssAssets();
+      $this->buildJsAssets($buildArgs);
+      $this->buildCssAssets($buildArgs);
     } else {
-      $this->moveFiles($this->sourceContext . $this->jsPath);
-      $this->moveFiles($this->sourceContext . $this->cssPath);
+      $this->copyDirectoryFilesIfNewer($this->sourceContext . $this->jsPath);
+      $this->copyDirectoryFilesIfNewer($this->sourceContext . $this->cssPath);
     }
 
     $this->copyDirectoryFilesIfNewer($this->sourceContext . $this->imagesPath);
 
-    $buildArgs['assets'] = array_merge(
-      $this->getBuiltAssets($this->sourceContext . $this->jsPath),
-      $this->getBuiltAssets($this->sourceContext . $this->cssPath)
+    return array_merge($buildArgs, ['assets' => $this->getBuiltAssets()]);
+  }
+
+  private function getBuiltAssets($from="") :array{
+    return $this->buildContext->fetchBuildFiles(
+      $this->sourceContext,
+      '/^.*(\.css)|(\.js)$/i'
     );
-    return $buildArgs;
   }
 
-  private function getBuiltAssets($from){
-    $files = [];
-    $build = $this->buildContext->getPath('build');
-    $directory = new \RecursiveDirectoryIterator($build . "/$from",
-      \RecursiveDirectoryIterator::SKIP_DOTS);
-    $iterator = new \RecursiveIteratorIterator($directory);
-    foreach(new \RecursiveIteratorIterator($directory) as $file){
-      $files[] = str_replace($build, "", "$file");
-    }
-    return $files;
-  }
-
-  private function buildJsAssets(){
+  private function buildJsAssets(array $buildArgs){
     switch($this->compression){
       case 'low':
         $jsFilters = [new JsMinifyFilter()];
@@ -129,10 +102,13 @@ class AssetBuilder implements BuilderInterface {
     foreach($jsFiles as $file){
       $jsCollection->add(new AssetCache(new FileAsset($file), $this->getCache()));
     }
-    file_put_contents($this->setupPath($this->sourceContext . $this->siteJsFile), $jsCollection->dump());
+    $this->buildContext->putBuildFile(
+      $this->sourceContext . $this->jsPath . "/{$buildArgs['buildID']}-site.js",
+      $jsCollection->dump()
+    );
   }
 
-  private function buildCssAssets(){
+  private function buildCssAssets(array $buildArgs){
     if($this->compression !== '') $cssFilters = [new CssMinifyFilter()];
     else $cssFilters = [];
 
@@ -148,7 +124,11 @@ class AssetBuilder implements BuilderInterface {
     foreach($cssFiles as $file){
       $cssCollection->add(new AssetCache(new FileAsset($file), $this->getCache()));
     }
-    file_put_contents($this->setupPath($this->sourceContext . $this->siteCssFile), $cssCollection->dump());
+
+    $this->buildContext->putBuildFile(
+      $this->sourceContext . $this->cssPath . "/{$buildArgs['buildID']}-site.css",
+      $cssCollection->dump()
+    );
   }
 
   private function copyDirectoryFilesIfNewer($dir){
@@ -162,34 +142,6 @@ class AssetBuilder implements BuilderInterface {
 
   private function isFileNewer($src, $dest) :bool{
     return (file_exists($dest) && filemtime($src) > filemtime($dest)) || !file_exists($dest);
-  }
-
-  private function moveFiles($dir){
-    // is copy good enough for this or should it be stream based?
-    if(is_dir($this->buildContext->getPath('source') . '/' . $dir)) {
-      $files = $this->buildContext->fetchSourceFiles($dir);
-      foreach($files as $file){
-        $this->moveFile($file);
-      }
-    }
-  }
-
-  private function moveFile($file){
-    $dest = $this->setupPath(str_replace($this->buildContext->getPath('source'), "", $file));
-    if((file_exists($dest) && filemtime($file) > filemtime($dest)) || !file_exists($dest)) {
-      copy($file, $dest);
-    }
-  }
-
-  private function setupPath($path){
-    $destination = $this->buildContext->getPath('build') .
-      DIRECTORY_SEPARATOR .
-      $path;
-
-    if (!file_exists(dirname($destination))) {
-      mkdir(dirname($destination), 0777, true);
-    }
-    return $destination;
   }
 
   private function getCache() :FilesystemCache {
